+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v1.2.3
+
+- Added caching for firebase commands
+  - Reduces the loading time to almost nothing
+
 v1.2.2
 
 - Added logic to cache command arguments